The great change
Moved the install on the laptop, meaning I had to resetup Github for the new instal directory. Also, 8.1 changed the file structure, so Addons have been moved to different folders. Finally, because I use my PC in the bedroom less, I've moved the curse client to the laptop, so I can update addons from there instead. Overall, this is a pretty massive change and I pray it will propagate correctly.